mechforge_ai/pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "mechforge-ai"
version = "0.3.0"
description = "MechForge AI - Mechanical Engineering AI Assistant"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}
authors = [
    {name = "MechForge Team"}
]
dependencies = [
    "pyyaml>=6.0",
    "rich>=13.0",
    "llama-cpp-python>=0.2.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-cov>=4.0",
]

[project.scripts]
mechforge = "mechforge_ai.main_cli:main"
mechforge-k = "mechforge_ai.knowledge_cli:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["mechforge_ai*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py", "*_test.py"]
python_functions = ["test_*"]
addopts = "-v --tb=short"
